requirements.txt
fastapi
uvicorn
openai
pdfplumber
requests
streamlit
PyPDF2
python-dotenv